chore: consolidate all SWR hooks
